# application/docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  v1.AccountIdBody:
    properties:
      accountId:
        type: string
    type: object
  v1.AccountRequestBody:
    properties:
      args:
        items:
          $ref: '#/definitions/v1.AccountIdBody'
        type: array
    type: object
  v1.Crop:
    properties:
      datatype:
        description: add column here
        type: string
      hashinfo:
        type: string
      id:
        type: string
    type: object
  v1.DonatingListQueryByGranteeRequestBody:
    properties:
      grantee:
        type: string
    type: object
  v1.DonatingListQueryRequestBody:
    properties:
      donor:
        type: string
    type: object
  v1.DonatingRequestBody:
    properties:
      donor:
        description: 捐赠人
        type: string
      grantee:
        description: 受赠人
        type: string
      objectOfDonating:
        description: 捐赠对象
        type: string
    type: object
  v1.RealEstateQueryRequestBody:
    properties:
      proprietor:
        description: 所有者(业主)(业主AccountId)
        type: string
    type: object
  v1.RealEstateRequestBody:
    properties:
      accountId:
        description: 操作人ID
        type: string
      livingSpace:
        description: 生活空间
        type: number
      proprietor:
        description: 所有者(业主)(业主AccountId)
        type: string
      totalArea:
        description: 总面积
        type: number
    type: object
  v1.SellingByBuyRequestBody:
    properties:
      buyer:
        description: 买家(买家AccountId)
        type: string
      objectOfSale:
        description: 销售对象(正在出售的房地产RealEstateID)
        type: string
      seller:
        description: 发起销售人、卖家(卖家AccountId)
        type: string
    type: object
  v1.SellingListQueryByBuyRequestBody:
    properties:
      buyer:
        description: 买家(买家AccountId)
        type: string
    type: object
  v1.SellingListQueryRequestBody:
    properties:
      seller:
        description: 发起销售人、卖家(卖家AccountId)
        type: string
    type: object
  v1.SellingRequestBody:
    properties:
      objectOfSale:
        description: 销售对象(正在出售的房地产RealEstateID)
        type: string
      price:
        description: 价格
        type: number
      salePeriod:
        description: 智能合约的有效期(单位为天)
        type: integer
      seller:
        description: 发起销售人、卖家(卖家AccountId)
        type: string
    type: object
  v1.UpdateDonatingRequestBody:
    properties:
      donor:
        description: 捐赠人
        type: string
      grantee:
        description: 受赠人
        type: string
      objectOfDonating:
        description: 捐赠对象
        type: string
      status:
        description: 需要更改的状态
        type: string
    type: object
  v1.UpdateSellingRequestBody:
    properties:
      buyer:
        description: 买家(买家AccountId)
        type: string
      objectOfSale:
        description: 销售对象(正在出售的房地产RealEstateID)
        type: string
      seller:
        description: 发起销售人、卖家(卖家AccountId)
        type: string
      status:
        description: 需要更改的状态
        type: string
    type: object
info:
  contact:
    email: 1390167880@qq.com
    name: fmy
  description: 基于区块链技术的农产品溯源系统api文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 基于区块链技术的农产品溯源系统api文档
  version: "1.0"
paths:
  /api/v1/addCrop:
    post:
      parameters:
      - description: crop
        in: body
        name: Crop
        required: true
        schema:
          $ref: '#/definitions/v1.Crop'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: test post
  /api/v1/checkDataBlockHeight:
    get:
      parameters:
      - description: 区块高度
        in: query
        name: datablockheight
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 检查查询的区块高度是否符合要求
  /api/v1/createDonating:
    post:
      parameters:
      - description: donating
        in: body
        name: donating
        required: true
        schema:
          $ref: '#/definitions/v1.DonatingRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 发起捐赠
  /api/v1/createRealEstate:
    post:
      parameters:
      - description: realEstate
        in: body
        name: realEstate
        required: true
        schema:
          $ref: '#/definitions/v1.RealEstateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 新建房地产(管理员)
  /api/v1/createSelling:
    post:
      parameters:
      - description: selling
        in: body
        name: selling
        required: true
        schema:
          $ref: '#/definitions/v1.SellingRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 发起销售
  /api/v1/createSellingByBuy:
    post:
      parameters:
      - description: sellingByBuy
        in: body
        name: sellingByBuy
        required: true
        schema:
          $ref: '#/definitions/v1.SellingByBuyRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 买家购买
  /api/v1/getBlockInfo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 输出blockInfo
  /api/v1/getBlockInfoByBlockHeight:
    get:
      parameters:
      - description: 区块高度
        in: query
        name: height
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 输出blockInfo
  /api/v1/getMaxDataBlockHeight:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 输出最大数据区块高度
  /api/v1/hello:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 测试输出Hello
  /api/v1/queryAccountList:
    post:
      parameters:
      - description: account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/v1.AccountRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取账户信息
  /api/v1/queryDonatingList:
    post:
      parameters:
      - description: donatingListQuery
        in: body
        name: donatingListQuery
        required: true
        schema:
          $ref: '#/definitions/v1.DonatingListQueryRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 查询捐赠列表(可查询所有，也可根据发起捐赠人查询)
  /api/v1/queryDonatingListByGrantee:
    post:
      parameters:
      - description: donatingListQueryByGrantee
        in: body
        name: donatingListQueryByGrantee
        required: true
        schema:
          $ref: '#/definitions/v1.DonatingListQueryByGranteeRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 根据受赠人(受赠人AccountId)查询捐赠(受赠的)(供受赠人查询)
  /api/v1/queryRealEstateList:
    post:
      parameters:
      - description: realEstateQuery
        in: body
        name: realEstateQuery
        required: true
        schema:
          $ref: '#/definitions/v1.RealEstateQueryRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取房地产信息(空json{}可以查询所有，指定proprietor可以查询指定业主名下房产)
  /api/v1/querySellingList:
    post:
      parameters:
      - description: sellingListQuery
        in: body
        name: sellingListQuery
        required: true
        schema:
          $ref: '#/definitions/v1.SellingListQueryRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 查询销售(可查询所有，也可根据发起销售人查询)(发起的)
  /api/v1/querySellingListByBuyer:
    post:
      parameters:
      - description: sellingListQueryByBuy
        in: body
        name: sellingListQueryByBuy
        required: true
        schema:
          $ref: '#/definitions/v1.SellingListQueryByBuyRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 根据参与销售人、买家(买家AccountId)查询销售(参与的)
  /api/v1/updateDonating:
    post:
      parameters:
      - description: updateDonating
        in: body
        name: updateDonating
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateDonatingRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 更新捐赠状态（确认受赠、取消）
  /api/v1/updateSelling:
    post:
      parameters:
      - description: updateSelling
        in: body
        name: updateSelling
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateSellingRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 更新销售状态（买家确认、买卖家取消）Status取值为 完成"done"、取消"cancelled" 当处于销售中状态，卖家要取消时，buyer为""空
swagger: "2.0"